requirements.txt
Flask==3.0.3
gunicorn
Flask-SQLAlchemy
python-dotenv
psycopg2-binary